+++ lib/src/internal/table_layout_render_box.dart
@@ -273,21 +273,34 @@ class TableLayoutRenderBox<ROW> extends RenderBox
                     _layoutSettings.themeMetrics.columnDividerThickness,
                     lastRowWidgetMetrics.height1),
                 columnPaint);
-            context.canvas.drawRect(
-                Rect.fromLTWH(
-                    left,
-                    lastRowWidgetMetrics.top2,
-                    _layoutSettings.themeMetrics.columnDividerThickness,
-                    lastRowWidgetMetrics.height2),
-                columnPaint);
+            if (_theme.columnDividerFillHeight) {
+              context.canvas.drawRect(
+                  Rect.fromLTWH(
+                      left,
+                      lastRowWidgetMetrics.top2,
+                      _layoutSettings.themeMetrics.columnDividerThickness,
+                      lastRowWidgetMetrics.height2),
+                  columnPaint);
+            }
           } else {
-            context.canvas.drawRect(
-                Rect.fromLTWH(
-                    left,
-                    offset.dy + _layoutSettings.themeMetrics.headerHeight,
-                    _layoutSettings.themeMetrics.columnDividerThickness,
-                    _layoutSettings.cellsBounds.height),
-                columnPaint);
+            if (_theme.columnDividerFillHeight) {
+              context.canvas.drawRect(
+                  Rect.fromLTWH(
+                      left,
+                      offset.dy + _layoutSettings.themeMetrics.headerHeight,
+                      _layoutSettings.themeMetrics.columnDividerThickness,
+                      _layoutSettings.cellsBounds.height),
+                  columnPaint);
+            } else {
+              context.canvas.drawRect(
+                  Rect.fromLTWH(
+                      left,
+                      offset.dy + _layoutSettings.themeMetrics.headerHeight,
+                      _layoutSettings.themeMetrics.columnDividerThickness,
+                      _layoutSettings.visibleRowsLength *
+                          _layoutSettings.themeMetrics.rowHeight),
+                  columnPaint);
+            }
           }
         }
         context.canvas.restore();